fix(media): refactor lightbox to a declarative React Portal with focus management
Replaces the imperative document.body DOM node and cross-component
reference-counting hack with a single ImageLightbox component rendered
via ReactDOM.createPortal from the root Layout. DocItem/Layout and
BlogPostItem/Container no longer need to mount the hook themselves,
since Layout already wraps every page - removing the redundant
multi-owner lifecycle that previously required activeConsumers
bookkeeping.

Also adds focus management: the previously focused element is restored
on close, and Tab is trapped on the dialog's only focusable control
while open. Tightens the logo/avatar heuristic to match whole class
tokens instead of relying on \b boundaries against hyphenated names.
