+++ src/theme/BlogPostItem/Container/index.js
@@ -1,10 +1,8 @@
 import React from "react";
 import clsx from "clsx";
 import { useBlogPost } from "@docusaurus/plugin-content-blog/client";
-import useImageLightbox from "../../utils/useImageLightbox";
 
 export default function BlogPostItemContainer({ children, className }) {
-  useImageLightbox();
   const { isBlogPostPage } = useBlogPost();
 
   return (
