--- src/theme/utils/useImageLightbox.js
@@ -1,273 +1,0 @@
-import { useEffect } from "react";
-import ExecutionEnvironment from "@docusaurus/ExecutionEnvironment";
-
-const MARKDOWN_IMAGE_SCOPE =
-  ".theme-doc-markdown, .theme-blog-markdown, article.markdown, .markdown";
-
-const BLOCKED_SCOPE =
-  ".no-lightbox, .avatar, .table-of-contents, .pagination-nav, .navbar, .footer";
-
-const MERMAID_CONTAINER = ".docusaurus-mermaid-container";
-
-let activeConsumers = 0;
-
-function isImageHref(href = "") {
-  return /\.(png|jpe?g|webp|gif|avif|svg)(\?|#|$)/i.test(href);
-}
-
-function shouldOpenLightbox(image) {
-  if (!(image instanceof HTMLImageElement)) {
-    return false;
-  }
-
-  if (!image.closest(MARKDOWN_IMAGE_SCOPE)) {
-    return false;
-  }
-
-  if (image.closest(BLOCKED_SCOPE)) {
-    return false;
-  }
-
-  return Boolean(image.currentSrc || image.src);
-}
-
-function sanitizeSvgIds(clone) {
-  const uniqueSuffix = `-lb-${Math.random().toString(36).substring(2, 8)}`;
-  const allElements = [clone, ...Array.from(clone.querySelectorAll("*"))];
-  const idMap = new Map();
-
-  allElements.forEach((el) => {
-    const oldId = el.getAttribute("id");
-    if (oldId) {
-      const newId = `${oldId}${uniqueSuffix}`;
-      idMap.set(oldId, newId);
-      el.setAttribute("id", newId);
-    }
-  });
-
-  if (idMap.size > 0) {
-    allElements.forEach((el) => {
-      Array.from(el.attributes).forEach((attr) => {
-        let val = attr.value;
-        let changed = false;
-
-        idMap.forEach((newId, oldId) => {
-          if (val.includes(`#${oldId}`)) {
-            val = val.replace(new RegExp(`#${oldId}\\b`, "g"), `#${newId}`);
-            changed = true;
-          }
-
-          if (attr.name === "aria-labelledby" || attr.name === "aria-describedby") {
-            const tokens = val.split(/\s+/);
-            if (tokens.includes(oldId)) {
-              val = tokens.map((t) => (t === oldId ? newId : t)).join(" ");
-              changed = true;
-            }
-          }
-        });
-
-        if (changed) {
-          el.setAttribute(attr.name, val);
-        }
-      });
-    });
-  }
-}
-
-function ensureLightbox() {
-  let root = document.querySelector(".hami-lightbox");
-  if (root) {
-    return root;
-  }
-
-  root = document.createElement("div");
-  root.className = "hami-lightbox";
-  root.setAttribute("role", "dialog");
-  root.setAttribute("aria-modal", "true");
-  root.setAttribute("aria-label", "Image preview");
-  root.hidden = true;
-
-  const closeButton = document.createElement("button");
-  closeButton.type = "button";
-  closeButton.className = "hami-lightbox__close";
-  closeButton.setAttribute("aria-label", "Close image preview");
-  closeButton.textContent = "×";
-
-  const lightboxImage = document.createElement("img");
-  lightboxImage.className = "hami-lightbox__image";
-  lightboxImage.alt = "";
-
-  const svgHost = document.createElement("div");
-  svgHost.className = "hami-lightbox__svg";
-  svgHost.hidden = true;
-
-  const caption = document.createElement("p");
-  caption.className = "hami-lightbox__caption";
-
-  root.appendChild(closeButton);
-  root.appendChild(lightboxImage);
-  root.appendChild(svgHost);
-  root.appendChild(caption);
-  document.body.appendChild(root);
-
-  const close = () => {
-    root.hidden = true;
-    document.body.classList.remove("hami-lightbox-open");
-    lightboxImage.removeAttribute("src");
-    lightboxImage.alt = "";
-    lightboxImage.hidden = false;
-    svgHost.hidden = true;
-    svgHost.replaceChildren();
-    caption.textContent = "";
-    root.dataset.hasCaption = "false";
-  };
-
-  const setCaption = (captionText = "") => {
-    const normalizedCaption = captionText.trim();
-    caption.textContent = normalizedCaption;
-    caption.hidden = !normalizedCaption;
-    root.dataset.hasCaption = normalizedCaption ? "true" : "false";
-  };
-
-  root.addEventListener("click", (event) => {
-    if (event.target === root) {
-      close();
-    }
-  });
-
-  lightboxImage.addEventListener("click", close);
-  svgHost.addEventListener("click", close);
-  closeButton.addEventListener("click", close);
-
-  const handleKeyDown = (event) => {
-    if (event.key === "Escape" && !root.hidden) {
-      close();
-    }
-  };
-
-  document.addEventListener("keydown", handleKeyDown);
-  root.__hamiLightboxCleanup = () => {
-    document.removeEventListener("keydown", handleKeyDown);
-    document.body.classList.remove("hami-lightbox-open");
-    if (root.parentNode) {
-      root.parentNode.removeChild(root);
-    }
-  };
-
-  root.__hamiLightboxOpen = ({ src, alt, caption: captionText }) => {
-    lightboxImage.src = src;
-    lightboxImage.alt = alt || "";
-    lightboxImage.hidden = false;
-    svgHost.hidden = true;
-    svgHost.replaceChildren();
-    setCaption(captionText || alt || "");
-    root.hidden = false;
-    document.body.classList.add("hami-lightbox-open");
-  };
-
-  root.__hamiLightboxOpenSvg = (svg, captionText) => {
-    const clone = svg.cloneNode(true);
-    sanitizeSvgIds(clone);
-
-    clone.style.maxWidth = "none";
-    clone.style.width = "100%";
-    clone.style.height = "100%";
-    clone.removeAttribute("height");
-    clone.removeAttribute("width");
-    svgHost.replaceChildren(clone);
-    svgHost.hidden = false;
-    lightboxImage.hidden = true;
-    lightboxImage.removeAttribute("src");
-    setCaption(captionText || "");
-    root.hidden = false;
-    document.body.classList.add("hami-lightbox-open");
-  };
-
-  return root;
-}
-
-function getImageCaption(image) {
-  const figureCaption = image.closest("figure")?.querySelector("figcaption")?.textContent;
-
-  return (
-    figureCaption ||
-    image.getAttribute("aria-label") ||
-    image.getAttribute("title") ||
-    image.getAttribute("alt") ||
-    ""
-  );
-}
-
-function handleImageClick(event) {
-  const target = event.target;
-  if (!(target instanceof Element)) {
-    return;
-  }
-
-  const mermaidContainer = target.closest(MERMAID_CONTAINER);
-  if (
-    mermaidContainer &&
-    mermaidContainer.closest(MARKDOWN_IMAGE_SCOPE) &&
-    !mermaidContainer.closest(BLOCKED_SCOPE)
-  ) {
-    const svg = mermaidContainer.querySelector("svg");
-    if (svg) {
-      const figure = mermaidContainer.closest("figure.mermaid-figure");
-      const captionText = figure?.querySelector("figcaption")?.textContent || "";
-      const lightbox = ensureLightbox();
-      lightbox.__hamiLightboxOpenSvg(svg, captionText);
-      return;
-    }
-  }
-
-  const image = target instanceof HTMLImageElement ? target : target.closest("img");
-  if (!shouldOpenLightbox(image)) {
-    return;
-  }
-
-  const parentLink = image.closest("a");
-  if (parentLink) {
-    const href = parentLink.getAttribute("href") || "";
-    const sameAsImage =
-      href === image.currentSrc || href === image.src || href === image.getAttribute("src");
-    if (!sameAsImage && !isImageHref(href)) {
-      return;
-    }
-    event.preventDefault();
-  }
-
-  const lightbox = ensureLightbox();
-  const caption = getImageCaption(image);
-  lightbox.__hamiLightboxOpen({
-    src: image.currentSrc || image.src,
-    alt: image.alt || caption,
-    caption,
-  });
-}
-
-export default function useImageLightbox() {
-  useEffect(() => {
-    if (!ExecutionEnvironment.canUseDOM) {
-      return undefined;
-    }
-
-    activeConsumers++;
-    if (!window.__hamiLightboxInitialized) {
-      window.__hamiLightboxInitialized = true;
-      document.addEventListener("click", handleImageClick);
-    }
-
-    return () => {
-      activeConsumers--;
-      if (activeConsumers <= 0) {
-        activeConsumers = 0;
-        document.removeEventListener("click", handleImageClick);
-        const root = document.querySelector(".hami-lightbox");
-        if (root && typeof root.__hamiLightboxCleanup === "function") {
-          root.__hamiLightboxCleanup();
-        }
-        window.__hamiLightboxInitialized = false;
-      }
-    };
-  }, []);
-}
